pyproject.toml
[tool.poetry]
name = "freelance-exchange"
version = "0.1.1"
description = ""
authors = ["Qcsteeven <leva.demonov2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg = "~=3.1.18"
psycopg-binary = "~=3.1.17"
python-dotenv = "~=1.0.1"
multipledispatch = "^1.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
